requirements.txt
scipy
scikit-learn
pytorch
torchvision
torchsummary
matplotlib
azure-ai-ml
azure-identity
python-dotenv
